add Authorization

+++ Services/DotNetInterview.Services.Data/CommentsService.cs
@@ -26,11 +26,15 @@
             throw new NotImplementedException();
         }
 
-        public async Task<bool> Delete(string commentId)
+        public async Task<bool> Delete(string commentId, string currentUserId, bool isAdmin)
         {
+
             var comment = await this.commentsRepository.GetByIdWithDeletedAsync(commentId);
 
-            this.commentsRepository.Delete(comment);
+            if (isAdmin || comment.UserId == currentUserId)
+            {
+                this.commentsRepository.Delete(comment);
+            }
 
             return (await this.commentsRepository.SaveChangesAsync()) != 0 ? true : false;
         }
